+++ src/Services/Dominion/Actions/SpellActionService.php
@@ -713,6 +713,9 @@ class SpellActionService
             }
             $duration += $target->getTechPerkValue('enemy_spell_duration');
             $duration += $target->getSpellPerkValue('enemy_spell_duration');
+            if ($target->hero !== null && $target->hero->getPerkValue('enemy_spell_duration')) {
+                $duration += $target->hero->getPerkValue('enemy_spell_duration');
+            }
 
             $activeSpell = $target->spells->find($spell->id);